+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-##[TBD]
+## [1.2.4]
 * Expose extra deviceInfo
+* Ignore sign in status on account removal from ODSP cache #1541
+* Disable bitcode #1552
 
 ## [1.2.3]
 * Stop extra background tasks in the system webview case.